# openapi/pileofmedia-openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "pileofmedia",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.3.0"
  },
  "servers": [
    {
      "url": "https://pileofmedia.com/api/",
      "description": "Production server"
    },
    {
      "url": "https://Localhost:5000/api/",
      "description": "Localhost"
    }
  ],
  "paths": {
...
    "/books": {
      "get": {
        "tags": [
          "Books"
        ],
        "operationId": "get_items",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned books",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooksTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "operationId": "create_book",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateBook"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Book successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/books/{id}": {
      "get": {
        "tags": [
          "Books"
        ],
        "operationId": "get_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the item",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned book based on the id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Books"
        ],
        "operationId": "delete_book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the item to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Book successfully deleted"
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/characters": {
      "get": {
        "tags": [
          "Characters"
        ],
        "operationId": "get_items",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned characters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharactersTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Characters"
        ],
        "operationId": "create_item",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacter"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Character successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Character"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/characters/name/{name}": {
      "get": {
        "tags": [
          "Characters"
        ],
        "operationId": "get_by_name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the item to search for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned characters based on the name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharactersTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/characters/{id}": {
      "get": {
        "tags": [
          "Characters"
        ],
        "operationId": "get_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the item",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned character based on the id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Character"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Characters"
        ],
        "operationId": "delete_item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the item to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Character successfully deleted"
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/franchises": {
      "get": {
        "tags": [
          "Franchises"
        ],
        "operationId": "get_items",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned franchises",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FranchisesTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Franchises"
        ],
        "operationId": "create_item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFranchise"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Franchise successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Franchise"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/franchises/name/{name}": {
      "get": {
        "tags": [
          "Franchises"
        ],
        "operationId": "get_by_name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the item to search for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned franchises based on the name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FranchisesTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "get_items",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BooksTotal": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "CharactersTotal": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Character"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "CreateBook": {
        "type": "object",
        "required": [
          "book",
          "covers"
        ],
        "properties": {
          "book": {
            "$ref": "#/components/schemas/CreateBookData"
          },
          "covers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateImage"
            }
          }
        }
      },
...
      "CreateCharacter": {
        "type": "object",
        "required": [
          "character"
        ],
        "properties": {
          "character": {
            "$ref": "#/components/schemas/CreateCharacterData"
          },
          "image": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CreateImage"
              }
            ],
            "nullable": true
          }
        }
      },
...
      "CreateImage": {
        "type": "string",
        "format": "binary"
      },
...
    },
    "securitySchemes": {
      "user_token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}